requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aitviewer=1.13.0=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
blas=1.0=mkl
blis=0.7.11=pypi_0
blobfile=3.0.0=pypi_0
bps-torch=0.1=pypi_0
brotli-python=1.0.9=py38h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.4=hbcca054_0
catalogue=2.0.10=pypi_0
certifi=2024.7.4=pyhd8ed1ab_0
chamfer-distance=0.1=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
chumpy=0.71=pypi_0
click=8.1.7=pypi_0
clip=1.0=pypi_0
cloudpathlib=0.19.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
confection=0.1.5=pypi_0
contourpy=1.1.1=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.12.1=pypi_0
cymem=2.0.8=pypi_0
docker-pycreds=0.4.0=pypi_0
dotmap=1.3.30=pypi_0
easydict=1.13=pypi_0
einops=0.4.1=pypi_0
ema-pytorch=0.0.10=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.16.1=pypi_0
fonttools=4.53.1=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.9.0=pypi_0
ftfy=6.2.0=pypi_0
fvcore=0.1.5.post20210915=py38
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
glcontext=2.5.0=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
h5py=3.11.0=pypi_0
huggingface-hub=0.25.2=pypi_0
human-body-prior=2.2.2.0=pypi_0
hydra=2.5=pypi_0
idna=3.7=py38h06a4308_0
imageio=2.19.3=pypi_0
imageio-ffmpeg=0.4.7=pypi_0
imgui=2.0.0=pypi_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
iopath=0.1.10=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
langcodes=3.4.0=pypi_0
language-data=1.2.0=pypi_0
lazy-loader=0.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuv=1.48.0=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
loguru=0.7.2=pypi_0
lxml=5.3.0=pypi_0
lz4-c=1.9.4=h6a678d5_1
marisa-trie=1.2.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.6.3=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
moderngl=5.10.0=pypi_0
moderngl-window=2.4.6=pypi_0
mpmath=1.3.0=pypi_0
multipledispatch=1.0.0=pypi_0
murmurhash=1.0.10=pypi_0
natsort=8.4.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=pypi_0
ninja=1.11.1.1=pypi_0
nltk=3.9.1=pypi_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
nvidiacub=1.10.0=0
omegaconf=2.3.0=pypi_0
opencv-contrib-python-headless=4.10.0.84=pypi_0
opencv-python=4.10.0.84=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h9ca470c_2
openssl=3.0.14=h5eee18b_0
packaging=24.1=pypi_0
pandas=2.0.3=pypi_0
pathtools=0.1.2=pypi_0
pillow=10.4.0=py38h5eee18b_0
pip=24.0=py38h06a4308_0
plyfile=1.0.3=pypi_0
portalocker=2.10.1=pypi_0
preshed=3.0.9=pypi_0
promise=2.3=pypi_0
protobuf=3.20.3=pypi_0
psutil=6.0.0=pypi_0
pycryptodomex=3.20.0=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pyglet=2.0.17=pypi_0
pygments=2.18.0=pypi_0
pyopengl=3.1.0=pypi_0
pyparsing=3.1.2=pypi_0
pyqt5=5.15.11=pypi_0
pyqt5-qt5=5.15.14=pypi_0
pyqt5-sip=12.15.0=pypi_0
pyrr=0.10.3=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=1.11.0=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytorch3d=0.7.0=pypi_0
pytz=2024.2=pypi_0
pywavefront=1.3.3=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
quaternion=3.5.2.post4=pypi_0
readline=8.2=h5eee18b_0
regex=2024.7.24=pypi_0
requests=2.32.2=py38h06a4308_0
rich=13.7.1=pypi_0
roma=1.5.0=pypi_0
safetensors=0.4.5=pypi_0
scenepic=1.0.8=pypi_0
scikit-image=0.20.0=pypi_0
scikit-learn=1.3.2=pypi_0
scikit-video=1.1.11=pypi_0
scipy=1.9.1=pypi_0
sentry-sdk=2.10.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.5.1=py38h06a4308_0
shellingham=1.5.4=pypi_0
shortuuid=1.0.13=pypi_0
six=1.16.0=pypi_0
smart-open=7.0.4=pypi_0
smmap=5.0.1=pypi_0
smplkit=0.0.3=pypi_0
smplx=0.1.28=pypi_0
spacy=3.7.6=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlite=3.45.3=h5eee18b_0
srsly=2.4.8=pypi_0
sympy=1.13.3=pypi_0
tabulate=0.9.0=pyhd8ed1ab_1
tbb=2021.8.0=hdb19cb5_0
termcolor=2.4.0=pyhd8ed1ab_0
thinc=8.2.5=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2023.7.10=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.20.1=pypi_0
torchaudio=0.11.0=py38_cu113
torchgeometry=0.1.2=pypi_0
torchvision=0.12.0=py38_cu113
tqdm=4.66.4=pyhd8ed1ab_0
transformers=4.45.2=pypi_0
trimesh=3.23.5=pypi_0
triton=3.0.0=pypi_0
typer=0.12.5=pypi_0
typing_extensions=4.11.0=py38h06a4308_0
tzdata=2024.1=pypi_0
urllib3=2.2.2=py38h06a4308_0
usd-core=24.8=pypi_0
wandb=0.12.21=pypi_0
wasabi=1.1.3=pypi_0
wcwidth=0.2.13=pypi_0
weasel=0.4.1=pypi_0
websockets=12.0=pypi_0
wheel=0.43.0=py38h06a4308_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
yacs=0.1.8=pyhd8ed1ab_0
yaml=0.2.5=h7f98852_2
zipp=3.19.2=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
